added static imags for apartments

+++ AndroidAppBack/airbnb/src/main/java/com/dit/airbnb/service/ImageService.java
@@ -189,6 +189,7 @@ public class ImageService {
             if (image.isEmpty()) continue;
             Image imageDel = image.get();
             imageRepository.delete(imageDel);
+            if (imageDel.getPath().contains("static_apart_image")) continue;
             Files.delete(Paths.get("src/main/resources/static/images/" + imageDel.getPath()));
         }
     }
